Move schema to internal instead of pkg since it's temporary
